move subcommand definitions out of main.rs, into individual mods
