+++ scenarios/univ2ConfigTest.toml
@@ -140,7 +140,7 @@ args = [
      "10000000000000"
 ]
 
-## SPAM ########################################################################
+### SPAM
 
 [[spam]]
